The bug from last time does not seem to affect normal gameplay very much so we are not worrying about it yet. Added code to drive a single MaxBoticsMB1040 ultrasonic distance sensor, and then untested code that is supposed to drive an array of them, since they need the RX pin to be powered to start the chain.
Started working on a fusion localizer, which takes input from three different localizer and decides what is the best data to use based on estimated field position. Also this requires me to write a localizer for the distance sensors that takes the best guess of our position and uses that to decide whether we can trust the distance sensor data, and if so, outputs what they think our position is.

Also made odometry retractable in teleop. We will use this with our fusion localizers for going over the bars.