Oooookay this was a lot. First, I had some package problems with the roadrunner package and it should work good for everyone now but LMK if something is still fudgy with it.
The big thing in thsi commit is what is in theory (still untested) the combined sensor fusion localizer we have been talking about. OdoMechDistLocalizer.java is a localizer that attempts to fuse the new DistanceSensorArrayLocalizer.java and the mecanum localizer and odo localizers from roadrunner. These are not tested and the distance sensors aren't even fully wired yet, but in theory this should work. The one thing that will definitely need tuning is 4 variables that are offsets for the reads of the distance sensors in the distance sensor localizer. the idea is that the sensors are not directly in the middle of the robot and so they need some offset to accurately give the position of the middle of the robot but I have not tuned them yet. I also do not currently account for being too close to the walls which I think we can be currently but that comes down to either remounting sensors or just adding a bit more logic. Will update once I get some more data.